backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg-c==3.1.19
PyJWT==2.8.0
python-decouple==3.8
setuptools==69.5.1
sqlparse==0.5.0
typing_extensions==4.12.2
wheel==0.43.0
